MASSIVE update and so many things changed IDK what to write.
Made map much bigger and designed it to be more PACMAN like.
Changed the whole way the map is used.
Made CGPA system with double. Lots of work arounds for this one.
Especially displaying a double to 2 decimal precision on SFML was tough and needed a lot of googleing and SO :)
Did some other stuff too which I dont remember.
Changed FOOD/Grade eating system and made lots of grades.
Made gradeMonsters unpredictable and random.(Took a lot of time)
Introduced a life aka Retakes system.